+++ templates/parts/admin/welcome-wordpress.php
@@ -8,11 +8,18 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.5.3
  */
 
 
 
+
+
+// Requirements check
+
+	if ( ! class_exists( 'Reykjavik_Welcome' ) ) {
+		return;
+	}
 
 
 ?>